// src/app/core/types/reports/BundledRequestReport.ts
export interface BundledRequestReportIf {
    businessKey: string;
    categoryType: string;
    oldNIC: string;
    newNIC: string;
    passport: string;
    customerName: string;
    submissionDate: string;
    isCasa: boolean;
    isCard: boolean;
    isLoan: boolean;
    personalLoanType: string;
    personalLoanTypeDesc: string;
    rsdVerification: boolean;
    empVerification: boolean;
    refVerification: boolean;
    frmalertCheckedDate: string;
    receivingDate: string;
    eyeBallingCompletedDate: string;
    relationshipMakerCompletionDate: string;
    relationshipCheckerCompletionDate: string;
    dataEntryMarkerCompletionDate: string;
    dataEntryCheckerCompletionDate: string;
    aipRMCompletionDate: string;
    aipDocCheckCompletionDate: string;
    rejectResubmissionDate: string;
    firstApprover: string;
    approvedDate: string;
    appStatus: string;
    cardType: string;
    sourceType: string;
    cardLimit: number;
    approvedAmount: number;
    netAmount: number;
    secondApproveDecisionDate: string;
    secondApprover: string;
    dsr: string;
    pendDate: string;
    pendReasons: string[];
    rejectReasons: string[];
    dbr: number;
    totalUnsecuredExposure: number;
    grossIncome: number;
    mueMultiplier: number;
    uwApprovedLevel: string;
    approvedLevelDesc: string;
    salesChannel: string;
    staticDataChanged: boolean;
    accountNumber: string;
    relationShipID: string;
    scoreCardView: string;
    showScoreCard: boolean;
    bundleProcessId: number;
    processInstance: string;
}